Reorganize AlphaFold 3 modules into dedicated alphafold3 subpackage
- Move all AlphaFold 3 modules to src/beignet/nn/alphafold3/
- Extract AtomAttentionDecoder from SampleDiffusion into separate module
- Clean up diffusers integration with proper imports
- Simplify root nn/__init__.py to only export main AlphaFold3 class
- Add comprehensive alphafold3/__init__.py with all component exports
- Fix import paths throughout codebase for new package structure
