+++ src/beignet/nn/alphafold3/_atom_transformer.py
