+++ src/beignet/nn/alphafold3/_msa.py
