+++ src/beignet/nn/alphafold3/_pairformer_stack.py
