+++ src/beignet/nn/alphafold3/_relative_position_encoding.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from .. import one_hot
+from beignet import one_hot
 
 
 class RelativePositionEncoding(nn.Module):
