+++ src/beignet/nn/alphafold3/_sample_diffusion.py
@@ -2,13 +2,9 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from beignet.nn import (
-    AtomAttentionEncoder,
-    AtomTransformer,
-    DiffusionTransformer,
-    RelativePositionEncoding,
-    Transition,
-)
+from . import AtomAttentionEncoder, DiffusionTransformer, RelativePositionEncoding
+from ._atom_attention_decoder import AtomAttentionDecoder
+from ._transition import Transition
 
 
 class SampleDiffusion(nn.Module):
@@ -34,7 +30,7 @@ class SampleDiffusion(nn.Module):
     Examples
     --------
     >>> import torch
-    >>> from beignet.nn import SampleDiffusion
+    >>> from beignet.nn.alphafold3 import SampleDiffusion
     >>> batch_size, n_atoms, n_tokens = 2, 1000, 32
     >>> module = SampleDiffusion()
     >>>
@@ -188,22 +184,6 @@ class _CentreRandomAugmentation(nn.Module):
     ----------
     s_trans : float, default=1.0
         Translation scale in Ångstroms
-
-    Examples
-    --------
-    >>> import torch
-    >>> from beignet.nn import _CentreRandomAugmentation
-    >>> batch_size, n_atoms = 2, 1000
-    >>> module = _CentreRandomAugmentation()
-    >>>
-    >>> x_t = torch.randn(batch_size, n_atoms, 3)
-    >>> x_t_augmented = module(x_t)
-    >>> x_t_augmented.shape
-    torch.Size([2, 1000, 3])
-
-    References
-    ----------
-    .. [1] AlphaFold 3 Algorithm 19: CentreRandomAugmentation
     """
 
     def __init__(self, s_trans: float = 1.0):
@@ -369,7 +349,7 @@ class _Diffusion(nn.Module):
         self.layer_norm_tokens = nn.LayerNorm(c_token)
 
         # Step 7: Atom attention decoder
-        self.atom_attention_decoder = _AtomAttentionDecoder(
+        self.atom_attention_decoder = AtomAttentionDecoder(
             c_token=c_token, c_atom=c_atom, n_head=n_head
         )
 
@@ -669,123 +649,3 @@ class _FourierEmbedding(nn.Module):
         # Algorithm 22 Step 2: return cos(2π(tw + b))
         # Broadcasting: (..., 1) * (c,) + (c,) -> (..., c)
         return torch.cos(2 * torch.pi * (input * self.w + self.b))
-
-
-class _AtomAttentionDecoder(nn.Module):
-    r"""
-    Atom Attention Decoder for AlphaFold 3.
-
-    This module broadcasts per-token activations to per-atom activations,
-    applies cross attention transformer, and maps to position updates.
-    Implements Algorithm 6 exactly.
-
-    Parameters
-    ----------
-    c_token : int, default=768
-        Channel dimension for token representations
-    c_atom : int, default=128
-        Channel dimension for atom representations
-    n_block : int, default=3
-        Number of transformer blocks
-    n_head : int, default=4
-        Number of attention heads
-
-    Examples
-    --------
-    >>> import torch
-    >>> from beignet.nn import _AtomAttentionDecoder
-    >>> batch_size, n_tokens, n_atoms = 2, 32, 1000
-    >>> module = _AtomAttentionDecoder()
-    >>>
-    >>> a = torch.randn(batch_size, n_tokens, 768)  # Token representations
-    >>> q_skip = torch.randn(batch_size, n_atoms, 768)  # Query skip
-    >>> c_skip = torch.randn(batch_size, n_atoms, 128)  # Context skip
-    >>> p_skip = torch.randn(batch_size, n_atoms, n_atoms, 16)  # Pair skip
-    >>>
-    >>> r_update = module(a, q_skip, c_skip, p_skip)
-    >>> r_update.shape
-    torch.Size([2, 1000, 3])
-
-    References
-    ----------
-    .. [1] AlphaFold 3 Algorithm 6: Atom attention decoder
-    """
-
-    def __init__(
-        self, c_token: int = 768, c_atom: int = 128, n_block: int = 3, n_head: int = 4
-    ):
-        super().__init__()
-
-        self.c_token = c_token
-        self.c_atom = c_atom
-        self.n_block = n_block
-        self.n_head = n_head
-
-        # Step 1: Broadcast per-token activations to per-atom activations
-        self.token_to_atom_proj = nn.Linear(c_token, c_token, bias=False)
-
-        # Step 2: Cross attention transformer
-        self.atom_transformer = AtomTransformer(
-            n_block=n_block,
-            n_head=n_head,
-            c_q=c_token,  # Query dimension
-            c_kv=c_atom,  # Key-value dimension
-            c_pair=None,  # Will be inferred from p_skip
-        )
-
-        # Step 3: Map to position updates
-        self.position_proj = nn.Linear(c_token, 3, bias=False)
-        self.layer_norm = nn.LayerNorm(c_token)
-
-    def forward(
-        self, a: Tensor, q_skip: Tensor, c_skip: Tensor, p_skip: Tensor
-    ) -> Tensor:
-        r"""
-        Forward pass of Atom Attention Decoder.
-
-        Implements Algorithm 6 exactly:
-        1. q_l = LinearNoBias(a_tok_idx(l)) + q_l^skip
-        2. {q_l} = AtomTransformer({q_l}, {c_l^skip}, {p_lm^skip}, N_block=3, N_head=4)
-        3. r_l^update = LinearNoBias(LayerNorm(q_l))
-
-        Parameters
-        ----------
-        a : Tensor, shape=(batch_size, n_tokens, c_token)
-            Token-level representations
-        q_skip : Tensor, shape=(batch_size, n_atoms, c_token)
-            Query skip connection
-        c_skip : Tensor, shape=(batch_size, n_atoms, c_atom)
-            Context skip connection
-        p_skip : Tensor, shape=(batch_size, n_atoms, n_atoms, c_atompair)
-            Pair skip connection
-
-        Returns
-        -------
-        r_update : Tensor, shape=(batch_size, n_atoms, 3)
-            Position updates for atoms
-        """
-        batch_size, n_tokens, c_token = a.shape
-        n_atoms = q_skip.shape[1]
-
-        # Step 1: Broadcast per-token activations to per-atom activations and add skip connection
-        # q_l = LinearNoBias(a_tok_idx(l)) + q_l^skip
-
-        # Create token indices for each atom (simple broadcasting approach)
-        # For simplicity, we'll map atoms to tokens cyclically
-        token_indices = torch.arange(n_atoms, device=a.device) % n_tokens
-
-        # Get corresponding token activations for each atom
-        a_tok_idx = a[:, token_indices]  # (batch_size, n_atoms, c_token)
-
-        # Apply linear projection and add skip connection
-        q = self.token_to_atom_proj(a_tok_idx) + q_skip
-
-        # Step 2: Cross attention transformer
-        # {q_l} = AtomTransformer({q_l}, {c_l^skip}, {p_lm^skip}, N_block=3, N_head=4)
-        q = self.atom_transformer(q, c_skip, p_skip)
-
-        # Step 3: Map to positions update
-        # r_l^update = LinearNoBias(LayerNorm(q_l))
-        r_update = self.position_proj(self.layer_norm(q))
-
-        return r_update
