+++ src/beignet/nn/alphafold3/_transition.py
