+++ src/beignet/nn/alphafold3/_triangle_attention_ending_node.py
