+++ src/beignet/nn/alphafold3/_triangle_attention_starting_node.py
